Assingment 6 Admin/User roolit

+++ src/api/models/cat-model.js
@@ -19,34 +19,54 @@ const findCatById = async (id) => {
 
 const addCat = async (cat, file) => {
   const {cat_name, weight, owner, filename, birthdate} = cat;
-  const sql = `INSERT INTO cats (cat_name, weight, owner, filename, birthdate)
+
+  if (authUser.role === 'admin' || authUser.user_id === owner) {
+    const sql = `INSERT INTO cats (cat_name, weight, owner, filename, birthdate)
                VALUES (?, ?, ?, ?, ?)`;
-  const params = [cat_name, weight, owner, filename, birthdate];
-  const rows = await promisePool.execute(sql, params);
-  console.log('rows', rows);
-  if (rows[0].affectedRows === 0) {
-    return false;
+    const params = [cat_name, weight, owner, filename, birthdate];
+    const rows = await promisePool.execute(sql, params);
+    console.log('rows', rows);
+    if (rows[0].affectedRows === 0) {
+      return false;
+    }
+    return {cat_id: rows[0].insertId};
+  } else {
+    throw new Error('Unauthorized');
   }
-  return {cat_id: rows[0].insertId};
 };
 
-const modifyCat = async (cat, id) => {
-  const sql = promisePool.format(`UPDATE cats SET ? WHERE cat_id = ?`, [cat, id]);
-  const rows = await promisePool.execute(sql);
-  console.log('rows', rows);
-  if (rows[0].affectedRows === 0) {
-    return false;
+const modifyCat = async (cat, id, authUser) => {
+  if (authUser.role === 'admin' || authUser.user_id === cat.owner) {
+    const sql = promisePool.format(`UPDATE cats SET ? WHERE cat_id = ?`, [cat, id]);
+    const rows = await promisePool.execute(sql);
+    console.log('rows', rows);
+    if (rows[0].affectedRows === 0) {
+      return false;
+    }
+    return {message: 'success'};
   }
-  return {message: 'success'};
 };
 
-const removeCat = async (id) => {
-  const [rows] = await promisePool.execute('DELETE FROM cats WHERE cat_id = ?', [id]);
-  console.log('rows', rows);
-  if (rows.affectedRows === 0) {
-    return false;
+const removeCat = async (id, authUser) => {
+  if (authUser.role === 'admin' || authUser.user_id === id) {
+    const [rows] = await promisePool.execute('DELETE FROM cats WHERE cat_id = ?', [id]);
+    console.log('rows', rows);
+    if (rows.affectedRows === 0) {
+      return false;
+    }
+    return {message: 'success'};
   }
-  return {message: 'success'};
 };
 
-export {listAllCats, findCatById, addCat, modifyCat, removeCat};
+const checkCatAuthorization = async (cat_id, user) => {
+  const [rows] = await promisePool.execute('SELECT * FROM cats WHERE cat_id = ?', [cat_id]);
+  if (rows.length === 0) {
+    return false;
+  }
+  if (rows[0].owner !== user.user_id) {
+    return false;
+  }
+  return true;
+};
+
+export {listAllCats, findCatById, addCat, modifyCat, removeCat, checkCatAuthorization};
